refactor(helm test): move tests from V2BakeryControllerWithArtifactDownloaderTest to rosco-integration
The tests V2BakeryControllerWithArtifactDownloaderTest depend on an executable named helm3 being in the path, and an executable named helm being helm v2.  This isn't typically the case (anymore) on the local file system of developer machines.

Moving the tests to rosco-integration means using the helm executables from the just-built docker image, where these dependencies are satisfied.  As well, as we upgrade the versions of helm in rosco, we'll verify that override handling logic continues to function.
