+++ rosco-integration/src/test/resources/Chart.yaml
