+++ rosco-integration/src/test/resources/expected_template.yaml
