+++ rosco-integration/src/test/resources/foo.yaml
