+++ rosco-integration/src/test/resources/logback.xml
@@ -32,5 +32,6 @@
     <logger name="tc" level="INFO"/>
     <logger name="com.github.dockerjava" level="WARN"/>
     <logger name="com.github.dockerjava.zerodep.shaded.org.apache.hc.client5.http.wire" level="OFF"/>
+    <logger name="wiremock.org.eclipse.jetty" level="INFO"/>
 
 </configuration>
