+++ rosco-integration/src/test/resources/values_external.yaml
